refactor(06): extract shared backup table-formatter helpers
Move shortULID, timeAgo, timeAgoSince, and renderProgressBar out of
cmd/dfsctl/commands/store/metadata/backup/{format.go,job/show.go} into
a new internal/cli/backupfmt subpackage. The duplication existed because
the backup/job subpackage cannot import its parent backup without a
cycle; a shared subpackage breaks the cycle cleanly. humanSize stays
package-local to backup (single-file helper, not worth a package hop).
