+++ cmd/dfsctl/commands/store/metadata/backup/format.go
@@ -1,47 +1,7 @@
 // Package backup implements per-store backup management commands.
 package backup
 
-import (
-	"fmt"
-	"strings"
-	"time"
-)
-
-// shortULID returns the first 8 chars of a ULID followed by an ellipsis
-// (U+2026 "…") so table listings stay compact while remaining copy-paste
-// friendly for "dfsctl ... backup show" prefix-hunting in scripts. D-26.
-func shortULID(id string) string {
-	const prefixLen = 8
-	if len(id) <= prefixLen {
-		return id
-	}
-	return id[:prefixLen] + "\u2026"
-}
-
-// timeAgo renders a duration relative to t such as "30s ago", "3m ago",
-// "3h ago", or "2d ago". Used in table-mode rendering for D-26's CREATED
-// / STARTED columns — JSON/YAML modes surface the raw RFC3339 timestamp.
-func timeAgo(t time.Time) string {
-	return timeAgoSince(t, time.Now())
-}
-
-// timeAgoSince is the testable seam — callers can pin "now" deterministically.
-func timeAgoSince(t, now time.Time) string {
-	d := now.Sub(t)
-	if d < 0 {
-		d = 0
-	}
-	switch {
-	case d < time.Minute:
-		return fmt.Sprintf("%ds ago", int(d.Seconds()))
-	case d < time.Hour:
-		return fmt.Sprintf("%dm ago", int(d.Minutes()))
-	case d < 24*time.Hour:
-		return fmt.Sprintf("%dh ago", int(d.Hours()))
-	default:
-		return fmt.Sprintf("%dd ago", int(d.Hours()/24))
-	}
-}
+import "fmt"
 
 // humanSize renders a byte count using binary units ("1.0MB", "234KB",
 // "12B"). Matches the existing dfsctl convention of a single decimal place
@@ -57,23 +17,4 @@ func humanSize(b int64) string {
 		exp++
 	}
 	return fmt.Sprintf("%.1f%cB", float64(b)/float64(div), "KMGTPE"[exp])
-}
-
-// renderProgressBar renders a fixed-width 20-cell progress bar for D-47.
-// Caller is responsible for suppressing it in non-table modes (JSON/YAML
-// surfaces the numeric Progress field instead). Out-of-range inputs are
-// clamped rather than errored — this is a rendering helper, not input
-// validation.
-func renderProgressBar(pct int) string {
-	const width = 20
-	if pct < 0 {
-		pct = 0
-	}
-	if pct > 100 {
-		pct = 100
-	}
-	filled := pct * width / 100
-	return fmt.Sprintf("%d%%  [%s%s]", pct,
-		strings.Repeat("\u2593", filled),
-		strings.Repeat("\u2591", width-filled))
 }
